refactor(cifs-Plugin): DaemonSet approach
Previously dependencies were installed using ansible. This change
installs the required 'jq' & 'cifs-utils' components using the kickstart
file at initial boot and then uses a daemonSet approach to obtain the
plugin and place it on the node. Once complete, the pod shifts to a
sleep state to keep the daemonset healthy.