feat(cluster): New script to scale down apps dependent on external storage
Apps dependent on external storage need to be scaled down to prevent
data loss and corruption when there is an issue or maintenance with the
external storage device. This script allows us to locate all PV's that
leverage a common share address (set via $NAS_ADDRESS variable) and
scale those apps down to 0 while labeling the namespace accordingly while also
creating a sync window to prevent clashing with argocd.

scaling up deletes the sync window and then forces a sync for all
namespaces that were labeled before, forcing the deployments to then
respect the replica count provided in-repo